app/__init__.py
from dotenv import load_dotenv
from flask import Flask, jsonify

from app.database import init_db
from app.routes import register_routes


def create_app(init_database=True):
    load_dotenv()

    app = Flask(__name__)

    if init_database:
        init_db(app)

    from app import models  # noqa: F401 - registers models with Peewee

    register_routes(app)

    @app.route("/health")
    def health():
        return jsonify(status="ok")

    @app.errorhandler(404)
    def not_found(_error):
        return jsonify(error="Not found"), 404

    @app.errorhandler(500)
    def internal_server_error(_error):
        return jsonify(error="Internal server error"), 500

    return app
